+++ dsfhub/resource_data_source.go
@@ -1403,7 +1403,7 @@ func resourceDSFDataSourceRead(d *schema.ResourceData, m interface{}) error {
 	if dsfDataSourceReadResponse.Data.AssetData.ServerPort != nil {
 		var serverPort string
 		if serverPortNum, ok := dsfDataSourceReadResponse.Data.AssetData.ServerPort.(float64); ok {
-			serverPort = fmt.Sprintf("%d", float64(serverPortNum))
+			serverPort = fmt.Sprintf("%d", int(serverPortNum))
 		} else {
 			serverPort = dsfDataSourceReadResponse.Data.AssetData.ServerPort.(string)
 		}
